found and fixed a bug in generate_poa(), data transferred was uint16_t instead of SizeT. Changed SizeT to int32_t, fixed formatting issues.

+++ cudapoa/src/cudapoa_kernels.cuh
@@ -46,7 +46,7 @@ namespace claragenomics
 namespace cudapoa
 {
 
-typedef uint16_t SizeT;
+typedef int32_t SizeT;
 
 /**
  * @brief A struct to hold information about the sequences
